refactor(main.go): refactored gen3-agent/agent.go to /gen3-agent/helpers/agent-helpers.go
